# public/.well-known/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sunu Yoon - API Covoiturage Sénégal",
    "description": "API pour rechercher, publier et réserver des covoiturages au Sénégal. Couvre toutes les villes: Dakar, Thiès, Saint-Louis, Touba, Kaolack, Ziguinchor, Mbour, Saly, etc. Prix en FCFA (XOF). Paiement Orange Money, Wave ou espèces.",
    "version": "1.0.0",
    "contact": {
      "name": "Sunu Yoon",
      "email": "support@sunuyoon.sn",
      "url": "https://sunuyoon.net"
    }
  },
  "servers": [
    {
      "url": "https://sunuyoon.net/api",
      "description": "Production"
    }
  ],
  "paths": {
    "/rides/recent": {
      "get": {
        "operationId": "getRecentRides",
        "summary": "Obtenir les trajets récents disponibles au Sénégal",
        "description": "Retourne les derniers trajets publiés avec places disponibles. Chaque trajet inclut: ville de départ, destination, date/heure, prix en FCFA, nombre de places, modèle de voiture, et informations du conducteur.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 10 },
            "description": "Nombre maximum de trajets à retourner"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des trajets récents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "rides": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Ride" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/rides/search": {
      "get": {
        "operationId": "searchRides",
        "summary": "Rechercher un covoiturage entre deux villes du Sénégal",
        "description": "Cherche des trajets disponibles par ville de départ, destination, date et nombre de places. Exemple: Dakar → Saint-Louis, Thiès → Mbour, Kaolack → Ziguinchor.",
        "parameters": [
          {
            "name": "origin",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Ville de départ (ex: Dakar, Thiès, Saint-Louis, Touba, Kaolack, Mbour)",
            "required": false
          },
          {
            "name": "destination",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Ville de destination (ex: Saint-Louis, Ziguinchor, Saly, Tambacounda)",
            "required": false
          },
          {
            "name": "date",
            "in": "query",
            "schema": { "type": "string", "format": "date" },
            "description": "Date du trajet au format YYYY-MM-DD",
            "required": false
          },
          {
            "name": "seats",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1 },
            "description": "Nombre minimum de places requises",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Résultats de recherche",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "rides": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Ride" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/rides/{id}": {
      "get": {
        "operationId": "getRideDetails",
        "summary": "Détail d'un trajet spécifique",
        "description": "Retourne toutes les informations d'un trajet: conducteur, prix, places restantes, caractéristiques du véhicule.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Identifiant unique du trajet"
          }
        ],
        "responses": {
          "200": {
            "description": "Détail du trajet",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "ride": { "$ref": "#/components/schemas/Ride" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Ride": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "driver": {
            "type": "object",
            "properties": {
              "id": { "type": "string" },
              "name": { "type": "string" },
              "rating": { "type": "number" },
              "reviewCount": { "type": "integer" },
              "isVerified": { "type": "boolean" }
            }
          },
          "origin": { "type": "string", "description": "Ville de départ" },
          "destination": { "type": "string", "description": "Ville d'arrivée" },
          "departureTime": { "type": "string", "format": "date-time" },
          "price": { "type": "number", "description": "Prix par place en FCFA (XOF)" },
          "currency": { "type": "string", "enum": ["XOF"] },
          "seatsAvailable": { "type": "integer" },
          "totalSeats": { "type": "integer" },
          "carModel": { "type": "string" },
          "description": { "type": "string" },
          "features": { "type": "array", "items": { "type": "string" } },
          "estimatedDuration": { "type": "string" }
        }
      },
...
    }
  }
}